core: Define different types of providers
Modify the core code to allow for a hooking provider.  This
is in addition to the existing core and utility providers.
Update checks to explicitly check for the desired type of
provider, rather than assuming that !util is core.

Hooking providers are identified by lacking a getinfo API
as part of their interface.  This prepares the code, so that
the existing hooks can be converted into providers, plus allow
for the possibility of external hooks.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>