prov/verbs: Verbs provider support for the InfiniBand XRC based MSG_EP.
Add support for the InfinBand XRC transport on XRC capable hardware. For
devices that support XRC, a second domain with the suffix "-xrc" is created
that supports FI_PROTO_RDMA_CM_IB_XRC.

The XRC implementation is intended be used as the core provider layered
with RxM. It requires that FI_OFI_RXM_USE_SRX=1 is set so that shared
receive contexts are used.

The Verbs provider environment variable FI_VERBS_PREFER_XRC=1 can be set
to prioritize XRC fi_info over RC fi_info, moving the XRC fi_info in front
of the RC fi_info. If not set, the RC fi_info will have priority.

The fi_info domain name can be used to explicitly select XRC or RC, and
will override the setting of FI_VERBS_PREFER_XRC (e.g. mlx5_0 or mlx5_0-xrc).

Both XRC INI and TGT QP are shared between message endpoints. XRC INI QP are
owned by the domain and shared between MSG endpoints within a process that
are sending to endpoints on the same remote node. XRC TGT QP are owned
by the linux kernel and are shared across processes by endpoints receiving
data from endpoints within the same remote process. XRC and RC can
not share the same domain. Depending on the extent of the sharing, it
may be desirable to increase the Verbs TX work queue size.

The RDMA CM is used for connection management setup only and creates only
the physical INI/TGT QP. Once connections are established, the associated
RDMA CM ID is destroyed. A subsequent UD method will be added for sending
disconnect notifications.

NOTE: The use of this code requires a small 3 line patch that permits
The RDMA CM to be used with connected QP without managing the QP as was
intended.

Signed-off-by: Steve Welch <swelch@systemfabricworks.com>