prov/verbs: Allow XRC transport to be selectable by info domain/protocol.
This change allows selection of the XRC/RC protocol to be done via the domain
specified in the fi_info. Getinfo will return two MSG_EP infos for each
device, with the latter domain name having the suffix "-xrc" and having
a protocol of FI_PROTO_RDMA_CM_IB_XRC.

The Verbs environment variable, FI_VERBS_PREFER_XRC, can be set to control
the ordering of the MSG_EP infos, giving the XRC info higher priority by
moving it in front of the RC info.  By default, the RC MSG_EP info will have
higher priority and should be selected as it previously was.

A domain specified in the hints will override the FI_VERBS_PERFER_XRC
setting, with the domain specified in the hints being used. (e.g. if
if a domain of "mlx5_0" is specified then RC will be used, or if a domain
of "mlx5_0-xrc" is specified then XRC will be used.

Ultimately it would be preferable to be be able to pass the protocol
(FI_PROTO_RDMA_CM_IB_RC or FI_PROTO_RDMA_CM_IB_XRC) through RxM to the
core provider (Verbs) to set which transport should be used.

Signed-off-by: Steve Welch <swelch@systemfabricworks.com>