prov/util: Add config options to disable monitors
This commit introduces two new configure options for controlling
how memory monitors are used with applications by default:

  --disable-uffd-monitor
  --disable-memhooks-monitor

When a monitor is disabled, it cannot be selected as the default monitor
at init, or when the environment variables are read in. If a user selects a
monitor with FI_MR_CACHE_MONITOR, then the following behavior is expected:
  - if the monitor is available, it becomes the default.
  - otherwise, no monitor is chosen as the default monitor and caching is disabled

The behavior choice was made so that users will experience a predictable behavior
when the monitor of choice is not available.

The behavior for default cache selection has been modified. If an available monitor is
disabled, the next monitor in order (memhooks, uffd, none) is selected if enabled.
If none of the enabled monitors are available, it defaults to none.
This change maintains the current preference order of monitors.

Signed-off-by: James Swaro <james.swaro@hpe.com>