contrib/cray: CUDA support for Shasta builds
Update the Shasta libfabric DST build process to install the CUDA
runtime development package which is required to build libfabric with
CUDA support enabled.

Libfabric CUDA support is configured with the dlopen argument which
allows libfabric to perform runtime checking to see if various CUDA
libraries are present. This allows libfabric to work with and without
these CUDA libraries being present.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>