contrib/cray: ROCR support for Shasta builds
Update the Shasta libfabric DST build process to install the ROCR
runtime development package which is required to build libfabric with
ROCR support enabled.

Libfabric ROCR support is configured with the dlopen argument which
allows libfabric to perform runtime checking to see if various ROCR
libraries are present. This allows libfabric to work with and without
these ROCR libraries being present.

ROCR RPMs are not provided for SLES 15 SP2. Disable ROCR for these
builds.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>